requirements.txt
os 
tempfile
streamlit
fitz
PIL
pytesseract
python-docx
joblib